refactor(checker): split qualified-name type analysis
