+++ crates/tsz-checker/src/state/type_analysis/core.rs
@@ -6,7 +6,6 @@ use crate::query_boundaries::checkers::generic as generic_query;
 use crate::query_boundaries::common::{lazy_def_id, type_param_info};
 use crate::state::CheckerState;
 use crate::symbol_resolver::TypeSymbolResolution;
-use crate::symbols_domain::alias_cycle::AliasCycleTracker;
 use rustc_hash::FxHashSet;
 use tracing::trace;
 use tsz_binder::{SymbolId, symbol_flags};
@@ -479,963 +478,6 @@ impl<'a> CheckerState<'a> {
         }
 
         all_updates
-    }
-
-    /// Resolve a qualified name (A.B.C) to its type.
-    ///
-    /// This function handles qualified type names like `Namespace.SubType`, `Module.Interface`,
-    /// or deeply nested names like `A.B.C`. It resolves each segment and looks up the final member.
-    ///
-    /// ## Resolution Strategy:
-    /// 1. **Recursively resolve left side**: For `A.B.C`, first resolve `A.B`
-    /// 2. **Get member type**: Look up rightmost member in left type's exports
-    /// 3. **Handle symbol merging**: Supports merged class+namespace, enum+namespace, etc.
-    ///
-    /// ## Qualified Name Forms:
-    /// - `Module.Type` - Type from module
-    /// - `Namespace.Interface` - Interface from namespace
-    /// - `A.B.C` - Deeply nested qualified name
-    /// - `Class.StaticMember` - Static class member
-    ///
-    /// ## Symbol Resolution:
-    /// - Checks exports of left side's symbol
-    /// - Handles merged symbols (class+namespace, function+namespace)
-    /// - Falls back to property access if not found in exports
-    ///
-    /// ## Error Reporting:
-    /// - TS2694: Namespace has no exported member
-    /// - Returns ERROR type if resolution fails
-    ///
-    /// ## Lib Binders:
-    /// - Collects lib binders for cross-arena symbol lookup
-    /// - Fixes TS2694 false positives for lib.d.ts types
-    ///
-    /// ## TypeScript Examples:
-    /// ```typescript
-    /// // Module members
-    /// namespace Utils {
-    ///   export interface Helper {}
-    /// }
-    /// let h: Utils.Helper;  // resolve_qualified_name("Utils.Helper")
-    ///
-    /// // Deep nesting
-    /// namespace A {
-    ///   export namespace B {
-    ///     export interface C {}
-    ///   }
-    /// }
-    /// let x: A.B.C;  // resolve_qualified_name("A.B.C")
-    ///
-    /// // Static class members
-    /// class Container {
-    ///   static class Inner {}
-    /// }
-    /// let y: Container.Inner;  // resolve_qualified_name("Container.Inner")
-    ///
-    /// // Merged symbols
-    /// function Model() {}
-    /// namespace Model {
-    ///   export interface Options {}
-    /// }
-    /// let opts: Model.Options;  // resolve_qualified_name("Model.Options")
-    /// ```
-    pub(crate) fn resolve_qualified_name(&mut self, idx: NodeIndex) -> TypeId {
-        let Some(node) = self.ctx.arena.get(idx) else {
-            return TypeId::ERROR; // Missing node - propagate error
-        };
-
-        let Some(qn) = self.ctx.arena.get_qualified_name(node) else {
-            return TypeId::ERROR; // Missing qualified name data - propagate error
-        };
-
-        let right_name = if let Some(right_node) = self.ctx.arena.get(qn.right) {
-            if let Some(id) = self.ctx.arena.get_identifier(right_node) {
-                id.escaped_text.clone()
-            } else {
-                return TypeId::ERROR; // Missing identifier data - propagate error
-            }
-        } else {
-            return TypeId::ERROR; // Missing right node - propagate error
-        };
-
-        // Resolve the left side (could be Identifier or another QualifiedName)
-        let left_type = if let Some(left_node) = self.ctx.arena.get(qn.left) {
-            let left_name = self.entity_name_text(qn.left).unwrap_or_default();
-
-            let sym_res = if left_node.kind == syntax_kind_ext::QUALIFIED_NAME {
-                self.resolve_qualified_symbol_in_type_position(qn.left)
-            } else if left_node.kind == SyntaxKind::Identifier as u16 {
-                let mut initial = self
-                    .resolve_identifier_symbol_as_qualified_type_anchor(qn.left)
-                    .map(TypeSymbolResolution::Type)
-                    .unwrap_or_else(|| self.resolve_identifier_symbol_in_type_position(qn.left));
-
-                // UMD `export as namespace Foo` makes Foo available as both a
-                // value and a type-position namespace anchor. The type-position
-                // resolver doesn't see UMD globals, so a bare `Foo.Thing`
-                // reference would otherwise emit TS2503 ("Cannot find namespace")
-                // even though the runtime/value side resolves fine. Promote to
-                // Type(umd_sym_id) so the existing namespace-member machinery
-                // downstream emits the correct TS2694 ("no exported member")
-                // for missing members.
-                if matches!(
-                    initial,
-                    TypeSymbolResolution::ValueOnly(_) | TypeSymbolResolution::NotFound
-                ) && !left_name.is_empty()
-                    && let Some(umd_sym_id) =
-                        self.resolve_umd_global_symbol_by_name(left_name.as_str())
-                {
-                    initial = TypeSymbolResolution::Type(umd_sym_id);
-                }
-
-                // When the left side of a qualified name resolves to a type parameter,
-                // it cannot serve as a namespace (no exports). In tsc, type parameters
-                // do NOT shadow namespace imports in qualified name positions like
-                // `E.Whatever` — the import `* as E` takes precedence over the type
-                // parameter `E`.  Fall back to file_locals lookup which finds imports
-                // and top-level declarations without type parameter shadowing.
-                let lib_binders = self.get_lib_binders();
-                if let TypeSymbolResolution::Type(sym_id) = initial {
-                    if self
-                        .ctx
-                        .binder
-                        .get_symbol_with_libs(sym_id, &lib_binders)
-                        .is_some_and(|s| s.has_any_flags(symbol_flags::TYPE_PARAMETER))
-                    {
-                        if let Some(file_sym) = self.ctx.binder.file_locals.get(&left_name) {
-                            let mut visited = AliasCycleTracker::new();
-                            let resolved = self
-                                .resolve_alias_symbol(file_sym, &mut visited)
-                                .unwrap_or(file_sym);
-                            if self
-                                .ctx
-                                .binder
-                                .get_symbol_with_libs(resolved, &lib_binders)
-                                .is_none_or(|s| !s.has_any_flags(symbol_flags::TYPE_PARAMETER))
-                            {
-                                TypeSymbolResolution::Type(resolved)
-                            } else {
-                                initial
-                            }
-                        } else {
-                            initial
-                        }
-                    } else {
-                        initial
-                    }
-                } else {
-                    initial
-                }
-            } else {
-                TypeSymbolResolution::NotFound
-            };
-
-            match sym_res {
-                TypeSymbolResolution::Type(sym_id) => {
-                    let lib_binders = self.get_lib_binders();
-                    if let Some(symbol) = self.ctx.binder.get_symbol_with_libs(sym_id, &lib_binders)
-                    {
-                        let valid_namespace_flags = symbol_flags::MODULE
-                            | symbol_flags::NAMESPACE_MODULE
-                            | symbol_flags::VALUE_MODULE
-                            | symbol_flags::CLASS
-                            | symbol_flags::ENUM
-                            | symbol_flags::REGULAR_ENUM
-                            | symbol_flags::CONST_ENUM
-                            | symbol_flags::ENUM_MEMBER;
-
-                        // Skip TS2713 for ALIAS symbols (imports) - they may target
-                        // namespaces in other files. Also skip when the resolved
-                        // export has an alias_partner (TYPE_ALIAS+ALIAS merge), as
-                        // the partner provides namespace access. Skip when parse
-                        // errors exist, as the qualified name may be malformed.
-                        let is_alias = symbol.has_any_flags(symbol_flags::ALIAS);
-                        let has_alias_partner =
-                            self.ctx.alias_partners_contains(self.ctx.binder, sym_id)
-                                || self.ctx.binder.resolve_import_symbol(sym_id).is_some_and(
-                                    |resolved| {
-                                        self.ctx.alias_partners_contains(self.ctx.binder, resolved)
-                                    },
-                                );
-                        if !symbol.has_any_flags(valid_namespace_flags)
-                            && !is_alias
-                            && !has_alias_partner
-                            && !self.ctx.has_parse_errors
-                        {
-                            let right_name = if let Some(right_node) = self.ctx.arena.get(qn.right)
-                                && let Some(id) = self.ctx.arena.get_identifier(right_node)
-                            {
-                                id.escaped_text.clone()
-                            } else {
-                                String::new()
-                            };
-
-                            // Get rightmost name of the left side
-                            let left_rightmost_name = if left_node.kind
-                                == syntax_kind_ext::QUALIFIED_NAME
-                            {
-                                if let Some(left_qn) = self.ctx.arena.get_qualified_name(left_node)
-                                {
-                                    if let Some(rn) = self.ctx.arena.get(left_qn.right)
-                                        && let Some(id) = self.ctx.arena.get_identifier(rn)
-                                    {
-                                        id.escaped_text.clone()
-                                    } else {
-                                        left_name.clone()
-                                    }
-                                } else {
-                                    left_name.clone()
-                                }
-                            } else {
-                                left_name.clone()
-                            };
-
-                            // Determine whether to emit TS2713 or TS2702.
-                            // TS2713: the property exists on the type — suggest indexed access.
-                            // TS2702: the property does NOT exist — generic "used as namespace" error.
-                            //
-                            // For type parameters, get_type_of_symbol may not return the
-                            // TypeParameter type. Check the type_parameter_scope first.
-                            let left_type_id = self
-                                .ctx
-                                .type_parameter_scope
-                                .get(&left_name)
-                                .copied()
-                                .unwrap_or_else(|| self.get_type_of_symbol(sym_id));
-                            let prop_exists =
-                                crate::query_boundaries::property_access::type_has_property(
-                                    self.ctx.types,
-                                    left_type_id,
-                                    &right_name,
-                                );
-
-                            use crate::diagnostics::diagnostic_codes;
-                            if prop_exists {
-                                self.error_at_node_msg(
-                                    idx, // The entire qualified name node
-                                    diagnostic_codes::CANNOT_ACCESS_BECAUSE_IS_A_TYPE_BUT_NOT_A_NAMESPACE_DID_YOU_MEAN_TO_RETRIEVE_THE,
-                                    &[left_rightmost_name.as_str(), right_name.as_str()],
-                                );
-                            } else {
-                                self.error_type_used_as_namespace_at(&left_rightmost_name, qn.left);
-                            }
-                            return TypeId::ERROR;
-                        }
-                    }
-                    self.type_reference_symbol_type(sym_id)
-                }
-                TypeSymbolResolution::ValueOnly(_) | TypeSymbolResolution::NotFound => {
-                    if left_node.kind == syntax_kind_ext::QUALIFIED_NAME {
-                        self.resolve_qualified_name(qn.left)
-                    } else if left_node.kind == SyntaxKind::Identifier as u16 {
-                        // globalThis is a synthetic namespace in TSC (flags = ValueModule | NamespaceModule)
-                        // with exports pointing to the global scope. Suppress TS2503 for it.
-                        if left_name == "globalThis" {
-                            return TypeId::ERROR;
-                        }
-                        if !self.is_unresolved_import_symbol(qn.left) && !left_name.is_empty() {
-                            // TS1212/TS1213/TS1214: when the lead identifier of a
-                            // qualified type name is a strict-mode reserved word
-                            // (`public.bar`, `private.x`, `interface.B`, ...) and we
-                            // are in strict mode, tsc emits the reserved-word error
-                            // alongside TS2503. The general TypeReference path covers
-                            // bare identifiers; qualified-name resolution lands here
-                            // and was missing the check.
-                            if crate::state_checking::is_strict_mode_reserved_name(
-                                left_name.as_str(),
-                            ) && self.is_strict_mode_for_node(qn.left)
-                            {
-                                self.emit_strict_mode_reserved_word_error_with_ast_walk(
-                                    qn.left,
-                                    left_name.as_str(),
-                                );
-                            }
-                            // Route through boundary for TS2503/TS2552 with suggestions
-                            let req = crate::query_boundaries::name_resolution::NameResolutionRequest::namespace(
-                                left_name.as_str(),
-                                qn.left,
-                            );
-                            match self.resolve_name_structured(&req) {
-                                Err(_failure) => {
-                                    // Emit TS2503 (cannot find namespace) with suggestions
-                                    self.error_cannot_find_namespace_with_suggestion(
-                                        left_name.as_str(),
-                                        qn.left,
-                                    );
-                                }
-                                Ok(_) => {
-                                    // Shouldn't happen since resolve_qualified_symbol_in_type_position
-                                    // already failed, but avoid false diagnostic
-                                }
-                            }
-                        }
-                        TypeId::ERROR
-                    } else {
-                        TypeId::ERROR
-                    }
-                }
-            }
-        } else {
-            TypeId::ERROR // Missing left node - propagate error
-        };
-
-        if left_type == TypeId::ANY || left_type == TypeId::ERROR {
-            return TypeId::ERROR; // Propagate error from left side
-        }
-
-        // Collect lib binders for cross-arena symbol lookup (fixes TS2694 false positives)
-        let lib_binders = self.get_lib_binders();
-
-        // First, try to resolve the left side as a symbol and check its exports.
-        // This handles merged class+namespace, function+namespace, and enum+namespace symbols.
-        let mut left_sym_for_missing = None;
-        let mut left_module_specifier: Option<String> = None;
-        let member_sym_id_from_symbol = if let Some(left_node) = self.ctx.arena.get(qn.left)
-            && left_node.kind == SyntaxKind::Identifier as u16
-        {
-            if let Some(sym_id) = self.resolve_identifier_symbol_as_qualified_type_anchor(qn.left) {
-                if let Some(symbol) = self.ctx.binder.get_symbol_with_libs(sym_id, &lib_binders) {
-                    left_sym_for_missing = Some(sym_id);
-                    left_module_specifier = symbol.import_module.clone();
-                    let mut result = self.resolve_symbol_export_for(
-                        Some(sym_id),
-                        symbol,
-                        &right_name,
-                        &lib_binders,
-                    );
-                    // TYPE_ALIAS+ALIAS merge: look up alias_partner and
-                    // resolve the member through the ALIAS symbol's namespace
-                    if result.is_none() {
-                        let alias_id =
-                            self.ctx
-                                .alias_partner_for(self.ctx.binder, sym_id)
-                                .or_else(|| {
-                                    let resolved = self.ctx.binder.resolve_import_symbol(sym_id)?;
-                                    self.ctx.alias_partner_for(self.ctx.binder, resolved)
-                                });
-                        if let Some(alias_id) = alias_id
-                            && let Some(alias_sym) =
-                                self.ctx.binder.get_symbol_with_libs(alias_id, &lib_binders)
-                        {
-                            result = alias_sym
-                                .exports
-                                .as_ref()
-                                .and_then(|exports| exports.get(&right_name));
-                            if result.is_none()
-                                && let Some(module) = alias_sym.import_module.as_ref()
-                            {
-                                // Resolve from ALIAS's source file, then
-                                // fall back to current-file resolution.
-                                result = self
-                                    .ctx
-                                    .resolve_alias_import_member(alias_id, module, &right_name)
-                                    .or_else(|| {
-                                        self.resolve_effective_module_exports(module)
-                                            .and_then(|e| e.get(&right_name))
-                                    });
-                            }
-                        }
-                    }
-                    if result.is_none() && self.entity_name_text(qn.left).as_deref() == Some("JSX")
-                    {
-                        result = self.resolve_jsx_namespace_merged_export_symbol_id(&right_name);
-                    }
-                    if result.is_none() {
-                        result = self.resolve_named_class_expression_namespace_member(
-                            qn.left,
-                            sym_id,
-                            &right_name,
-                        );
-                    }
-                    result
-                } else {
-                    None
-                }
-            } else {
-                None
-            }
-        } else {
-            // When the left side is a QualifiedName (e.g., `ns.Root` in `ns.Root.Foo`),
-            // extract the module specifier from the root identifier of the chain so that
-            // module augmentation merging can be applied to nested members.
-            if left_module_specifier.is_none()
-                && let Some(left_node) = self.ctx.arena.get(qn.left)
-                && left_node.kind == syntax_kind_ext::QUALIFIED_NAME
-            {
-                left_module_specifier = self.extract_root_module_specifier(qn.left, &lib_binders);
-            }
-            None
-        };
-
-        // If found via symbol resolution, use it
-        if let Some(member_sym_id) = member_sym_id_from_symbol {
-            if (self.alias_resolves_to_value_only(member_sym_id, Some(right_name.as_str()))
-                || self.symbol_is_value_only(member_sym_id, Some(right_name.as_str())))
-                && !self.symbol_is_type_only(member_sym_id, Some(right_name.as_str()))
-            {
-                let full_name = self
-                    .entity_name_text(idx)
-                    .unwrap_or_else(|| right_name.clone());
-                self.report_wrong_meaning(
-                    &full_name,
-                    idx,
-                    member_sym_id,
-                    crate::query_boundaries::name_resolution::NameLookupKind::Value,
-                    crate::query_boundaries::name_resolution::NameLookupKind::Type,
-                );
-                return TypeId::ERROR;
-            }
-            let mut member_type = self.type_reference_symbol_type(member_sym_id);
-            if let Some(module_specifier) = left_module_specifier.as_deref() {
-                member_type =
-                    self.apply_module_augmentations(module_specifier, &right_name, member_type);
-            }
-            return member_type;
-        }
-
-        if let Some(left_sym_id) = left_sym_for_missing
-            && let Some(symbol) = self
-                .ctx
-                .binder
-                .get_symbol_with_libs(left_sym_id, &lib_binders)
-            && (symbol.flags
-                & (symbol_flags::MODULE
-                    | symbol_flags::CLASS
-                    | symbol_flags::REGULAR_ENUM
-                    | symbol_flags::CONST_ENUM
-                    | symbol_flags::INTERFACE)
-                != 0
-                // UMD `export as namespace Foo` produces an ALIAS symbol whose
-                // import_module points at the source file. tsc treats Foo as a
-                // namespace anchor in type position, so a missing member emits
-                // TS2694 ("no exported member") rather than TS2503 ("cannot
-                // find namespace"). Accept the UMD alias here so the
-                // member-resolution failure below routes through the TS2694
-                // branch.
-                || symbol.is_umd_export)
-        {
-            // If the left symbol is a pure interface (no namespace meaning) and a
-            // local declaration shadows an outer namespace, the member might exist
-            // on the outer namespace. In tsc, import-equals and qualified type names
-            // prefer namespace meaning, so a local interface shouldn't cause TS2694
-            // when an outer namespace with the same name has the member.
-            let is_pure_interface = symbol.has_any_flags(symbol_flags::INTERFACE)
-                && !symbol.has_any_flags(symbol_flags::MODULE)
-                && !symbol.has_any_flags(symbol_flags::CLASS)
-                && !symbol.has_any_flags(symbol_flags::REGULAR_ENUM)
-                && !symbol.has_any_flags(symbol_flags::CONST_ENUM);
-            if is_pure_interface {
-                // Check if an outer namespace with this name has the member
-                let left_name_str = self
-                    .entity_name_text(qn.left)
-                    .unwrap_or_else(|| symbol.escaped_name.clone());
-                if self
-                    .resolve_outer_namespace_member(qn.left, &left_name_str, &right_name)
-                    .is_some()
-                {
-                    // The member exists on an outer namespace — don't emit TS2694.
-                    // Return ERROR type since we can't resolve through the local interface,
-                    // but avoid the misleading diagnostic.
-                    return TypeId::ERROR;
-                }
-            }
-
-            let export_names: Vec<String> = symbol
-                .exports
-                .as_ref()
-                .map(|e| e.iter().map(|(name, _)| name.clone()).collect())
-                .unwrap_or_default();
-            let req =
-                crate::query_boundaries::name_resolution::NameResolutionRequest::exported_member(
-                    &right_name,
-                    qn.right,
-                    left_sym_id,
-                    export_names,
-                );
-            let failure = match self.resolve_name_structured(&req) {
-                Err(f) => f,
-                Ok(_) => {
-                    // Shouldn't happen since we already failed above, but be safe
-                    return TypeId::ERROR;
-                }
-            };
-            self.report_name_resolution_failure(&req, &failure);
-            return TypeId::ERROR;
-        }
-
-        // Otherwise, fall back to type-based lookup for pure namespace/module types
-        // Look up the member in the left side's exports
-        // Supports both Lazy(DefId) and Enum types
-        let fallback_sym_id = self.ctx.resolve_type_to_symbol_id(left_type);
-
-        if let Some(fallback_sym) = fallback_sym_id
-            && let Some(symbol) = self
-                .ctx
-                .binder
-                .get_symbol_with_libs(fallback_sym, &lib_binders)
-        {
-            // Use the helper to resolve the member from exports, members, or re-exports
-            if let Some(member_sym_id) = self.resolve_symbol_export_for(
-                Some(fallback_sym),
-                symbol,
-                &right_name,
-                &lib_binders,
-            ) {
-                // Check value-only, but skip for namespaces since they can be used
-                // to navigate to types (e.g., Outer.Inner.Type)
-                if let Some(member_symbol) = self
-                    .ctx
-                    .binder
-                    .get_symbol_with_libs(member_sym_id, &lib_binders)
-                {
-                    let is_namespace = member_symbol.has_any_flags(symbol_flags::MODULE);
-                    if !is_namespace
-                        && (self
-                            .alias_resolves_to_value_only(member_sym_id, Some(right_name.as_str()))
-                            || self.symbol_is_value_only(member_sym_id, Some(right_name.as_str())))
-                        && !self.symbol_is_type_only(member_sym_id, Some(right_name.as_str()))
-                    {
-                        let full_name = self
-                            .entity_name_text(idx)
-                            .unwrap_or_else(|| right_name.clone());
-                        self.report_wrong_meaning(
-                            &full_name,
-                            idx,
-                            member_sym_id,
-                            crate::query_boundaries::name_resolution::NameLookupKind::Value,
-                            crate::query_boundaries::name_resolution::NameLookupKind::Type,
-                        );
-                        return TypeId::ERROR;
-                    }
-                }
-                let mut member_type = self.type_reference_symbol_type(member_sym_id);
-                if let Some(module_specifier) = left_module_specifier.as_deref() {
-                    member_type =
-                        self.apply_module_augmentations(module_specifier, &right_name, member_type);
-                }
-                return member_type;
-            }
-
-            // If the symbol is an enum member, it cannot have exports.
-            // Emit TS2749 (value used as type) instead of TS2694.
-            if symbol.has_any_flags(symbol_flags::ENUM_MEMBER) {
-                let full_name = self
-                    .entity_name_text(idx)
-                    .unwrap_or_else(|| right_name.clone());
-                self.report_wrong_meaning(
-                    &full_name,
-                    idx,
-                    fallback_sym,
-                    crate::query_boundaries::name_resolution::NameLookupKind::Value,
-                    crate::query_boundaries::name_resolution::NameLookupKind::Type,
-                );
-                return TypeId::ERROR;
-            }
-
-            // Not found - report TS2694 or TS2724 (with spelling suggestion)
-            let export_names: Vec<String> = symbol
-                .exports
-                .as_ref()
-                .map(|e| e.iter().map(|(name, _)| name.clone()).collect())
-                .unwrap_or_default();
-            let req =
-                crate::query_boundaries::name_resolution::NameResolutionRequest::exported_member(
-                    &right_name,
-                    qn.right,
-                    fallback_sym,
-                    export_names,
-                );
-            let failure = match self.resolve_name_structured(&req) {
-                Err(f) => f,
-                Ok(_) => {
-                    return TypeId::ERROR;
-                }
-            };
-            self.report_name_resolution_failure(&req, &failure);
-            return TypeId::ERROR;
-        }
-
-        // Left side wasn't a reference to a namespace/module
-        // This is likely an error - the left side should resolve to a namespace
-        // Emit an appropriate error for the unresolved qualified name
-        // We don't emit TS2304 here because the left side might have already emitted an error
-        // Returning ERROR prevents cascading errors while still indicating failure
-        if let Some(left_node) = self.ctx.arena.get(qn.left)
-            && left_node.kind == SyntaxKind::Identifier as u16
-            && !self.is_unresolved_import_symbol(qn.left)
-            && let Some(ident) = self.ctx.arena.get_identifier(left_node)
-        {
-            self.error_cannot_find_namespace_with_suggestion(ident.escaped_text.as_str(), qn.left);
-        }
-        TypeId::ERROR
-    }
-
-    /// Walk a qualified-name chain leftward to find the root identifier and return
-    /// its `import_module` (module specifier), if any.  This is used to propagate
-    /// module augmentation context through nested qualified names like `ns.Root.Foo`.
-    pub(crate) fn extract_root_module_specifier(
-        &self,
-        mut idx: NodeIndex,
-        lib_binders: &[std::sync::Arc<tsz_binder::BinderState>],
-    ) -> Option<String> {
-        loop {
-            let node = self.ctx.arena.get(idx)?;
-            if node.kind == syntax_kind_ext::QUALIFIED_NAME {
-                let qn = self.ctx.arena.get_qualified_name(node)?;
-                idx = qn.left;
-            } else if node.kind == SyntaxKind::Identifier as u16 {
-                let sym_id = self.resolve_identifier_symbol_as_qualified_type_anchor(idx)?;
-                let symbol = self.ctx.binder.get_symbol_with_libs(sym_id, lib_binders)?;
-                return symbol.import_module.clone();
-            } else {
-                return None;
-            }
-        }
-    }
-
-    /// When a named class expression shadows an outer namespace of the same name,
-    /// qualified type names like `C.Member` still resolve through the namespace.
-    fn resolve_named_class_expression_namespace_member(
-        &self,
-        node: NodeIndex,
-        sym_id: SymbolId,
-        member_name: &str,
-    ) -> Option<SymbolId> {
-        let lib_binders = self.get_lib_binders();
-        let symbol = self.ctx.binder.get_symbol_with_libs(sym_id, &lib_binders)?;
-        if !symbol.has_any_flags(symbol_flags::CLASS) {
-            return None;
-        }
-
-        let decl_kind = self
-            .ctx
-            .arena
-            .get(symbol.value_declaration)
-            .map(|decl| decl.kind)?;
-        if decl_kind != syntax_kind_ext::CLASS_EXPRESSION {
-            return None;
-        }
-
-        self.resolve_outer_namespace_member(node, symbol.escaped_name.as_str(), member_name)
-    }
-
-    /// Resolve a member from an outer namespace with the same name.
-    /// Used to avoid false TS2694 when a local declaration shadows an outer namespace.
-    fn resolve_outer_namespace_member(
-        &self,
-        node: NodeIndex,
-        namespace_name: &str,
-        member_name: &str,
-    ) -> Option<SymbolId> {
-        let lib_binders = self.get_lib_binders();
-        // Walk up scopes from the enclosing scope's parent
-        let Some(scope_id) = self.ctx.binder.find_enclosing_scope(self.ctx.arena, node) else {
-            return self.resolve_namespace_member_from_all_binders(namespace_name, member_name);
-        };
-        let Some(current_scope) = self.ctx.binder.scopes.get(scope_id.0 as usize) else {
-            return self.resolve_namespace_member_from_all_binders(namespace_name, member_name);
-        };
-        let mut walk_id = current_scope.parent;
-
-        while let Some(scope) = self.ctx.binder.scopes.get(walk_id.0 as usize) {
-            if let Some(sym_id) = scope.table.get(namespace_name)
-                && let Some(sym) = self.ctx.binder.get_symbol_with_libs(sym_id, &lib_binders)
-                && sym.has_any_flags(symbol_flags::NAMESPACE)
-            {
-                // Found a namespace - check if it has the member
-                if let Some(exports) = sym.exports.as_ref()
-                    && let Some(member_id) = exports.get(member_name)
-                {
-                    return Some(member_id);
-                }
-            }
-            if walk_id == scope.parent {
-                break;
-            }
-            walk_id = scope.parent;
-        }
-
-        self.resolve_namespace_member_from_all_binders(namespace_name, member_name)
-    }
-
-    /// Resolve a member from a symbol's exports, members, or re-exports.
-    ///
-    /// This helper implements the common pattern of looking up a member in:
-    /// 1. Direct exports
-    /// 2. Members (for classes with static members)
-    /// 3. Re-exports (for imported namespaces)
-    ///
-    /// Returns `Some(member_sym_id)` if found, `None` otherwise.
-    fn resolve_symbol_export_for(
-        &mut self,
-        sym_id: Option<tsz_binder::SymbolId>,
-        symbol: &tsz_binder::Symbol,
-        member_name: &str,
-        lib_binders: &[std::sync::Arc<tsz_binder::BinderState>],
-    ) -> Option<tsz_binder::SymbolId> {
-        // Try direct exports first
-        if let Some(ref exports) = symbol.exports
-            && let Some(member_id) = exports.get(member_name)
-        {
-            return Some(member_id);
-        }
-
-        // For classes, also check members (for static members in type queries)
-        // This handles `typeof C.staticMember` where C is a class
-        if symbol.has_any_flags(symbol_flags::CLASS)
-            && let Some(ref members) = symbol.members
-            && let Some(member_id) = members.get(member_name)
-        {
-            return Some(member_id);
-        }
-
-        if symbol.has_any_flags(symbol_flags::MODULE) {
-            if let Some(member_id) =
-                self.resolve_module_export_from_declarations(symbol, member_name)
-            {
-                return Some(member_id);
-            }
-            // Only the anonymous source-file module should see top-level file exports
-            // through `file_locals`. Named namespaces/modules must resolve members
-            // through their own declaration-local export tables; otherwise `X.bar`
-            // can accidentally bind to an unrelated `export function bar()` from the
-            // containing file and surface TS2749 instead of TS2694.
-            let has_named_module_declaration = symbol.declarations.iter().any(|&decl_idx| {
-                let Some(node) = self.ctx.arena.get(decl_idx) else {
-                    return false;
-                };
-                if node.kind != syntax_kind_ext::MODULE_DECLARATION {
-                    return false;
-                }
-                self.ctx
-                    .arena
-                    .get_module(node)
-                    .and_then(|module| self.ctx.arena.get(module.name))
-                    .is_some_and(|name_node| name_node.kind == SyntaxKind::Identifier as u16)
-            });
-            if !has_named_module_declaration
-                && let Some(local_sym_id) = self.ctx.binder.file_locals.get(member_name)
-                && let Some(sym) = self.ctx.binder.get_symbol(local_sym_id)
-                && sym.is_exported
-            {
-                return Some(local_sym_id);
-            }
-        }
-
-        // If not found in direct exports, check for re-exports
-        // The member might be re-exported from another module
-        if let Some(ref module_specifier) = symbol.import_module {
-            if self
-                .ctx
-                .namespace_import_alias_has_local_namespace_conflict(symbol)
-            {
-                return None;
-            }
-            if symbol.has_any_flags(symbol_flags::ALIAS)
-                && self
-                    .ctx
-                    .module_resolves_to_non_module_entity(module_specifier)
-            {
-                return None;
-            }
-            if let Some(reexported_sym_id) =
-                self.resolve_reexported_member(module_specifier, member_name, lib_binders)
-            {
-                return Some(reexported_sym_id);
-            }
-            // Cross-file fallback: resolve the relative module specifier from
-            // the ALIAS symbol's source file perspective.
-            if let Some(alias_id) = sym_id
-                && let Some(resolved) =
-                    self.ctx
-                        .resolve_alias_import_member(alias_id, module_specifier, member_name)
-            {
-                return Some(resolved);
-            }
-        }
-
-        None
-    }
-
-    fn resolve_module_export_from_declarations(
-        &self,
-        symbol: &tsz_binder::Symbol,
-        member_name: &str,
-    ) -> Option<tsz_binder::SymbolId> {
-        for &decl_idx in &symbol.declarations {
-            let Some(node) = self.ctx.arena.get(decl_idx) else {
-                continue;
-            };
-            if node.kind != syntax_kind_ext::MODULE_DECLARATION {
-                continue;
-            }
-            let Some(module) = self.ctx.arena.get_module(node) else {
-                continue;
-            };
-            if module.body.is_none() {
-                continue;
-            }
-            if let Some(&scope_id) = self.ctx.binder.node_scope_ids.get(&module.body.0)
-                && let Some(scope) = self.ctx.binder.scopes.get(scope_id.0 as usize)
-                && let Some(sym_id) = scope.table.get(member_name)
-                && let Some(sym) = self.ctx.binder.get_symbol(sym_id)
-                && sym.is_exported
-            {
-                return Some(sym_id);
-            }
-            let Some(module_block) = self.ctx.arena.get_module_block_at(module.body) else {
-                continue;
-            };
-            let Some(statements) = &module_block.statements else {
-                continue;
-            };
-
-            for &stmt_idx in &statements.nodes {
-                let Some(stmt_node) = self.ctx.arena.get(stmt_idx) else {
-                    continue;
-                };
-                if (stmt_node.kind == syntax_kind_ext::TYPE_ALIAS_DECLARATION
-                    || stmt_node.kind == syntax_kind_ext::INTERFACE_DECLARATION)
-                    && let Some(name) = self.get_declaration_name_text(stmt_idx)
-                    && name == member_name
-                    && let Some(&sym_id) = self.ctx.binder.node_symbols.get(&stmt_idx.0)
-                {
-                    return Some(sym_id);
-                }
-                if stmt_node.kind != syntax_kind_ext::EXPORT_DECLARATION {
-                    continue;
-                }
-                let Some(export_decl) = self.ctx.arena.get_export_decl(stmt_node) else {
-                    continue;
-                };
-                if export_decl.export_clause.is_none() {
-                    continue;
-                }
-                let Some(clause_node) = self.ctx.arena.get(export_decl.export_clause) else {
-                    continue;
-                };
-
-                match clause_node.kind {
-                    syntax_kind_ext::FUNCTION_DECLARATION
-                    | syntax_kind_ext::CLASS_DECLARATION
-                    | syntax_kind_ext::INTERFACE_DECLARATION
-                    | syntax_kind_ext::TYPE_ALIAS_DECLARATION
-                    | syntax_kind_ext::ENUM_DECLARATION
-                    | syntax_kind_ext::MODULE_DECLARATION => {
-                        if let Some(name) =
-                            self.get_declaration_name_text(export_decl.export_clause)
-                            && name == member_name
-                            && let Some(&sym_id) = self
-                                .ctx
-                                .binder
-                                .node_symbols
-                                .get(&export_decl.export_clause.0)
-                        {
-                            return Some(sym_id);
-                        }
-                    }
-                    syntax_kind_ext::VARIABLE_STATEMENT => {
-                        if let Some(var_stmt) = self.ctx.arena.get_variable(clause_node) {
-                            // VariableStatement holds VariableDeclarationList nodes.
-                            // Walk list -> declaration to recover exported namespace vars.
-                            for &list_idx in &var_stmt.declarations.nodes {
-                                let Some(list_node) = self.ctx.arena.get(list_idx) else {
-                                    continue;
-                                };
-                                let Some(decl_list) = self.ctx.arena.get_variable(list_node) else {
-                                    continue;
-                                };
-                                for &decl_idx in &decl_list.declarations.nodes {
-                                    if let Some(name) = self.get_declaration_name_text(decl_idx)
-                                        && name == member_name
-                                        && let Some(&sym_id) =
-                                            self.ctx.binder.node_symbols.get(&decl_idx.0)
-                                    {
-                                        return Some(sym_id);
-                                    }
-                                }
-                            }
-                        }
-                    }
-                    // Handle `export { A }` and `export { A as B }` — named export
-                    // specifiers that re-export symbols from the enclosing scope.
-                    syntax_kind_ext::NAMED_EXPORTS => {
-                        let Some(named) = self.ctx.arena.get_named_imports(clause_node) else {
-                            continue;
-                        };
-                        for &spec_idx in &named.elements.nodes {
-                            let Some(spec_node) = self.ctx.arena.get(spec_idx) else {
-                                continue;
-                            };
-                            let Some(spec) = self.ctx.arena.get_specifier(spec_node) else {
-                                continue;
-                            };
-                            // The exported name: for `export { A }`, name is "A";
-                            // for `export { A as B }`, name is "B".
-                            let exported_name = self
-                                .ctx
-                                .arena
-                                .get(spec.name)
-                                .and_then(|n| self.ctx.arena.get_identifier(n))
-                                .map(|id| id.escaped_text.as_str());
-                            let Some(exported) = exported_name else {
-                                continue;
-                            };
-                            if exported != member_name {
-                                continue;
-                            }
-                            // The local name: for `export { A }`, local is "A";
-                            // for `export { A as B }`, local is "A" (property_name).
-                            let local_name_idx = if spec.property_name.is_some() {
-                                spec.property_name
-                            } else {
-                                spec.name
-                            };
-                            let local_name = self
-                                .ctx
-                                .arena
-                                .get(local_name_idx)
-                                .and_then(|n| self.ctx.arena.get_identifier(n))
-                                .map(|id| id.escaped_text.as_str());
-                            let Some(local) = local_name else {
-                                continue;
-                            };
-                            // Resolve the local name: try the module body scope first,
-                            // then walk up to parent scopes and file locals.
-                            if let Some(&scope_id) =
-                                self.ctx.binder.node_scope_ids.get(&module.body.0)
-                            {
-                                // Check the module scope itself
-                                if let Some(scope) = self.ctx.binder.scopes.get(scope_id.0 as usize)
-                                    && let Some(sym_id) = scope.table.get(local)
-                                {
-                                    return Some(sym_id);
-                                }
-                                // Walk parent scopes
-                                let mut current_scope_id = scope_id;
-                                while let Some(scope) =
-                                    self.ctx.binder.scopes.get(current_scope_id.0 as usize)
-                                {
-                                    if let Some(sym_id) = scope.table.get(local) {
-                                        return Some(sym_id);
-                                    }
-                                    if current_scope_id == scope.parent {
-                                        break; // root scope
-                                    }
-                                    current_scope_id = scope.parent;
-                                }
-                            }
-                            // Fall back to file locals
-                            if let Some(sym_id) = self.ctx.binder.file_locals.get(local) {
-                                return Some(sym_id);
-                            }
-                        }
-                    }
-                    _ => {}
-                }
-            }
-        }
-        None
     }
 
     /// Get type from a union type node (A | B).
