+++ crates/tsz-checker/src/state/type_analysis/mod.rs
@@ -35,6 +35,7 @@ mod cross_file_overlay_gate;
 pub(crate) mod cross_file_query_types;
 mod cross_file_residue;
 mod cross_file_shared_cache;
+mod qualified_names;
 mod source_alias_attribution;
 mod symbol_type_helpers;
 mod type_param_defaults;
